refactor(front): move tier catalog lookups to token_pricing, drop client mirror
