+++ front-api/routes/w/[wId]/model_tiers/index.ts
@@ -1,4 +1,4 @@
-import { listTiers } from "@app/lib/model_tiers/allowed_tiers";
+import { listTiers } from "@app/lib/api/assistant/token_pricing/tiers";
 import type { GetModelTiersResponseBody } from "@app/types/api/model_tiers";
 import { workspaceApp } from "@front-api/middlewares/ctx";
 import { ensureIsAdmin } from "@front-api/middlewares/ensure_role";
