+++ front/lib/api/assistant/token_pricing/tiers.ts
@@ -1,5 +1,6 @@
 import { STATIC_MODEL_SUPPORTED_REASONING_EFFORTS } from "@app/lib/api/assistant/token_pricing/static_model_reasoning_efforts";
 import type { StaticModelIdType } from "@app/types/assistant/models/models";
+import { isStaticModelId } from "@app/types/assistant/models/models";
 import type {
   ModelIdType,
   ModelProviderIdType,
@@ -464,3 +465,28 @@ export const STATIC_MODEL_TIERS: StaticModelTiersLookup = {
     none: "premium",
   },
 } satisfies StaticModelTiersMap;
+
+export function listTiers(): readonly ModelsTierDefinition[] {
+  return MODELS_TIERS;
+}
+
+export function getTier(name: ModelsTierName): ModelsTierDefinition | null {
+  return MODELS_TIERS.find((tier) => tier.name === name) ?? null;
+}
+
+export function getTierForSelection(
+  selection: ModelTierSelection
+): ModelsTierName | null {
+  return getTierForModel(selection.modelId, selection.reasoningEffort);
+}
+
+export function getTierForModel(
+  modelId: ModelTierSelection["modelId"],
+  reasoningEffort: ModelTierSelection["reasoningEffort"]
+): ModelsTierName | null {
+  // includes models added at runtime on GCP (EAPs)
+  if (!isStaticModelId(modelId)) {
+    return "premium";
+  }
+  return STATIC_MODEL_TIERS[modelId][reasoningEffort] ?? null;
+}
