+++ front/lib/model_tiers/enabled_models.ts
@@ -1,13 +1,13 @@
 import { pickPreferredLargeModel } from "@app/lib/api/assistant/model_preferences";
 import type { ModelsTierName } from "@app/lib/api/assistant/token_pricing/tiers";
-import { STATIC_MODEL_TIERS } from "@app/lib/api/assistant/token_pricing/tiers";
+import {
+  getTierForModel,
+  STATIC_MODEL_TIERS,
+} from "@app/lib/api/assistant/token_pricing/tiers";
 import { getAvailableModelsForWorkspace } from "@app/lib/api/assistant/workspace_capabilities";
 import type { Authenticator } from "@app/lib/auth";
 import { getFeatureFlags } from "@app/lib/auth";
-import {
-  getTierForModel,
-  resolveAllowedTierNames,
-} from "@app/lib/model_tiers/allowed_tiers";
+import { resolveAllowedTierNames } from "@app/lib/model_tiers/allowed_tiers";
 import type {
   EnabledModelConfigurationType,
   ModelStreamResolutionsType,
